bank: re-organize some of the packages
The infrastructure packages are now properly organized using the
orthogonal/ports & adapters architecture (GOOS).

bank.core
bank.adapter.http
